requirements.txt
ccxt>=4.0.0
pandas>=2.0.0
numpy>=1.24.0
pandas-ta>=0.3.14b
mplfinance>=0.12.10b0
matplotlib>=3.7.0
python-telegram-bot>=20.0
python-dotenv>=1.0.0
scipy>=1.10.0
aiohttp>=3.8.0
